requirements.txt
asgiref==3.8.1
babel==2.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
crispy-bootstrap5==2025.4
Django==5.2
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.19.1
idna==3.10
Jinja2==3.1.6
lxml==5.3.2
MarkupSafe==3.0.2
python-docx==1.1.2
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
gunicorn
PyPDF2==3.0.1
pdfplumber==0.11.4
